requirements.txt
flask
flask-cors
deepface
opencv-python
numpy
twilio
tf-keras
python-dotenv
